Correctly Remove PHP If There Are Multiple Versions
We now stop all instances of PHP from booting when removing PHP and rely
on the individual tasks to start PHP correctly.

If some how, there are multiple versions of PHP installed, we will loop
over all versions and remove the ones that are not needed.

This should make switching PHP versions while processes are still
running a "fixable" issue by re-running the provisioning script.